get contract address from db
