--- allocator-before-allocation/config.json
@@ -1,3 +1,0 @@
-{
-	"contractAddress": "0xb15E6CeAED167Bf5a50aB780BD0e758d96D44856"
-}
