--- lockup-lockedup/config.json
@@ -1,3 +1,0 @@
-{
-	"contractAddress": "0x3d40fab11ee30e3aa1900ccfafd190f0851a6157"
-}
