--- market-factory-create/config.json
@@ -1,3 +1,0 @@
-{
-	"contractAddress": "0x6fbb138ac2e9f780826ed10cc05211d9506d8ad9"
-}
