--- metrics-factory-create/config.json
@@ -1,3 +1,0 @@
-{
-	"contractAddress": "0xaf490ba3eaff9495f01407ec22027bd90eafabb1"
-}
