--- policy-factory-create/config.json
@@ -1,3 +1,0 @@
-{
-	"contractAddress": "0xca0f09564b1d0182b907352c631734d65c457d77"
-}
