+++ policy-factory-create/index.ts
@@ -3,7 +3,6 @@ import { AzureFunction, Context } from '@azure/functions'
 import { ObjectType } from 'typeorm'
 import { EventSaver } from '../common/base'
 import { PolicyFactoryCreate } from '../entities/policy-factory-create'
-import config from './config.json'
 import abi from './abi.json'
 
 class CreateEventSaver extends EventSaver {
@@ -23,10 +22,6 @@ class CreateEventSaver extends EventSaver {
 
 	getBatchName(): string {
 		return 'policy-factory-create'
-	}
-
-	getContractAddress(): string {
-		return config.contractAddress
 	}
 
 	// eslint-disable-next-line @typescript-eslint/no-untyped-public-signature
