+++ property-factory-create/index.ts
@@ -2,7 +2,6 @@ import { AzureFunction, Context } from '@azure/functions'
 import { ObjectType } from 'typeorm'
 import { EventSaver } from '../common/base'
 import { PropertyFactoryCreate } from '../entities/property-factory-create'
-import config from './config.json'
 import abi from './abi.json'
 
 class CreateEventSaver extends EventSaver {
@@ -22,10 +21,6 @@ class CreateEventSaver extends EventSaver {
 
 	getBatchName(): string {
 		return 'property-factory-create'
-	}
-
-	getContractAddress(): string {
-		return config.contractAddress
 	}
 
 	// eslint-disable-next-line @typescript-eslint/no-untyped-public-signature
